✨ feat: Refactor configuration schemas and types for improved structure and clarity
