+++ packages/data-provider/src/react-query/react-query-service.ts
@@ -5,6 +5,7 @@ import type {
   QueryObserverResult,
 } from '@tanstack/react-query';
 import { initialModelsConfig, LocalStorageKeys } from '../config';
+import type { TStartupConfig } from '../config';
 import { defaultOrderQuery } from '../types/assistants';
 import * as dataService from '../data-service';
 import * as m from '../types/mutations';
@@ -423,18 +424,14 @@ export const useUpdateUserPluginsMutation = (
 };
 
 export const useGetStartupConfig = (
-  config?: UseQueryOptions<t.TStartupConfig>,
-): QueryObserverResult<t.TStartupConfig> => {
-  return useQuery<t.TStartupConfig>(
-    [QueryKeys.startupConfig],
-    () => dataService.getStartupConfig(),
-    {
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      refetchOnMount: false,
-      ...config,
-    },
-  );
+  config?: UseQueryOptions<TStartupConfig>,
+): QueryObserverResult<TStartupConfig> => {
+  return useQuery<TStartupConfig>([QueryKeys.startupConfig], () => dataService.getStartupConfig(), {
+    refetchOnWindowFocus: false,
+    refetchOnReconnect: false,
+    refetchOnMount: false,
+    ...config,
+  });
 };
 
 export const useGetCustomConfigSpeechQuery = (
